feat: Update data types for achievement tracking and minting logic to support larger values

+++ src/achievements/systems.cairo
@@ -22,14 +22,14 @@ impl AchievementImpl<S, +WorldTrait<S>, +Drop<S>> of Achievements<S> {
     }
 
     fn _progress_achievement(
-        ref self: S, player_id: ContractAddress, task: TaskId, count: u32, time: u64,
+        ref self: S, player_id: ContractAddress, task: TaskId, count: u128, time: u64,
     ) {
         let mut storage = self.storage(ACHIEVEMENTS_NAMESPACE_HASH);
         storage.emit_achievement_progress(player_id.into(), task, count, time);
     }
 
     fn progress_achievement(
-        ref self: S, player_id: ContractAddress, task: TaskId, count: u32, time: u64,
+        ref self: S, player_id: ContractAddress, task: TaskId, count: u128, time: u64,
     ) {
         if count > 0 {
             self._progress_achievement(player_id, task, count, time);
@@ -37,7 +37,7 @@ impl AchievementImpl<S, +WorldTrait<S>, +Drop<S>> of Achievements<S> {
     }
 
     fn progress_achievements(
-        ref self: S, player_id: ContractAddress, task_and_counts: Array<(TaskId, u32)>, time: u64,
+        ref self: S, player_id: ContractAddress, task_and_counts: Array<(TaskId, u128)>, time: u64,
     ) {
         let mut storage = self.storage(ACHIEVEMENTS_NAMESPACE_HASH);
         let player_felt: felt252 = player_id.into();
@@ -49,13 +49,15 @@ impl AchievementImpl<S, +WorldTrait<S>, +Drop<S>> of Achievements<S> {
     }
 
     fn progress_achievements_now(
-        ref self: S, player_id: ContractAddress, task_and_counts: Array<(TaskId, u32)>,
+        ref self: S, player_id: ContractAddress, task_and_counts: Array<(TaskId, u128)>,
     ) {
         self.progress_achievements(player_id, task_and_counts, get_block_timestamp());
     }
 
 
-    fn progress_achievement_now(ref self: S, player_id: ContractAddress, task: TaskId, count: u32) {
+    fn progress_achievement_now(
+        ref self: S, player_id: ContractAddress, task: TaskId, count: u128,
+    ) {
         self.progress_achievement(player_id, task, count, get_block_timestamp());
     }
 
@@ -68,7 +70,7 @@ impl AchievementImpl<S, +WorldTrait<S>, +Drop<S>> of Achievements<S> {
     }
 
     fn increment_achievements(
-        ref self: S, player_id: ContractAddress, task_and_counts: Array<(TaskId, u32)>, time: u64,
+        ref self: S, player_id: ContractAddress, task_and_counts: Array<(TaskId, u128)>, time: u64,
     ) {
         let mut storage = self.storage(ACHIEVEMENTS_NAMESPACE_HASH);
         let player_felt: felt252 = player_id.into();
@@ -80,7 +82,7 @@ impl AchievementImpl<S, +WorldTrait<S>, +Drop<S>> of Achievements<S> {
     }
 
     fn increment_achievements_now(
-        ref self: S, player_id: ContractAddress, task_and_counts: Array<(TaskId, u32)>,
+        ref self: S, player_id: ContractAddress, task_and_counts: Array<(TaskId, u128)>,
     ) {
         self.increment_achievements(player_id, task_and_counts, get_block_timestamp());
     }
